Start writing some tests for this new implementation of system messages
