+++ Rocket.Chat/Managers/Model/MessageManager/MessageManagerSystemMessages.swift
@@ -1,0 +1,35 @@
+//
+//  MessageManagerSystemMessages.swift
+//  Rocket.Chat
+//
+//  Created by Rafael Kellermann Streit on 10/09/18.
+//  Copyright © 2018 Rocket.Chat. All rights reserved.
+//
+
+import Foundation
+import SwiftyJSON
+import RealmSwift
+
+extension MessageManager {
+
+    static func createSystemMessage(from object: [String: JSON], realm: Realm? = Realm.current) {
+        guard let realm = realm else { return }
+        guard let subscriptionIdentifier = object["rid"]?.string else { return }
+        guard let detachedSubscription = Subscription.find(rid: subscriptionIdentifier, realm: realm) else { return }
+
+        realm.execute({ (realm) in
+            let message = Message.getOrCreate(realm: realm, values: JSON(object), updates: { (object) in
+                object?.subscription = detachedSubscription
+            })
+
+            if message.user == nil {
+                let user = User.getOrCreate(realm: realm, values: ["id": "rocket.cat", "username": "rocket.cat"], updates: nil)
+                message.user = user
+            }
+
+            message.privateMessage = true
+            realm.add(message, update: true)
+        })
+    }
+
+}
