+++ Rocket.ChatTests/Managers/MessageManagerSpec.swift
@@ -209,4 +209,7 @@ extension MessageManagerSpec {
             XCTAssert(messages.count == 0, "fetchMessages() will not return userMuted or userUnmuted when they're hidden")
         })
     }
+
+    func te
+
 }
